# internal/routers/api/v1/tag.go
package v1

import (
	"github.com/gin-gonic/gin"
	"github.com/go-programming-tour-book/blog-service/global"
	"github.com/go-programming-tour-book/blog-service/pkg/app"
	"github.com/go-programming-tour-book/blog-service/pkg/errcode"
)

type Tag struct {}

func NewTag() Tag{
	return Tag{}
}

//	@Summary 	获取单个标签
//	@Produce 	json
//	@Param		id				body		int			false		"标签ID"
//	@Success	200		{object}		model.TagSwagger	"成功"
//	@Failure	400		{object}		errcode.Error		"请求错误"
//	@Failure	500		{object}		errcode.Error		"内部错误"
//	@Router		/api/v1/tags/{id}	[get]
func (t Tag) Get (c *gin.Context){}

//	@Summary 	获取多个标签
//	@Produce 	json
//	@Param		name		query		string		false		"标签名称"		maxlength(100)
//	@Param		state		query		int			false		"状态"			Enums(0, 1)	default(1)
//	@Param		page		query		int			false		"页码"
//	@Param		page_size	query		int			false		"每页数量"
//	@Success	200		{object}		model.TagSwagger	"成功"
//	@Failure	400		{object}		errcode.Error		"请求错误"
//	@Failure	500		{object}		errcode.Error		"内部错误"
//	@Router		/api/v1/tags	[get]
func (t Tag) List (c *gin.Context){
	//c.JSON(200, gin.H{"detail" : global.DatabaseSetting.Password})
	//returnInfo := errcode.ServerError
	//if returnInfo.StatusCode() != http.StatusOK {
	//	global.Logger.Info("http response was error")
	//}
	//app.NewResponse(c).ToErrorResponse(errcode.ServerError)

	param := struct {
		Name 		string		`form:"name"				binding:"max=100"`
		State 		uint8		`form:"state,default=1"		binding:"oneof=0 1"`
	}{}
	response := app.NewResponse(c)
	valid, errs := app.BindAndValid(c, &param)
	if !valid {
		global.Logger.Errorf("app.BindAndValid errs: %v", errs)
		errRsp := errcode.InvalidParams.WithDetails(errs.Errors()...)
		response.ToErrorResponse(errRsp)
		return
	}

	response.ToResponse(gin.H{})
	return
}

//	@Summary 	新增标签
//	@Produce 	json
//	@Param		name		body		string		false		"标签名称"		minlength(3) maxlength(100)
//	@Param		state		body		int			false		"状态"			Enums(0, 1)	default(1)
//	@Param		created_by	body		string		false		"创建者"		minlength(3) maxlength(100)
//	@Success	200		{object}		model.TagSwagger	"成功"
//	@Failure	400		{object}		errcode.Error		"请求错误"
//	@Failure	500		{object}		errcode.Error		"内部错误"
//	@Router		/api/v1/tags	[post]
func (t Tag) Create (c *gin.Context){}

//	@Summary 	更新标签
//	@Produce 	json
//	@Param		id				body		int			false		"标签ID"
//	@Param		name			body		string		false		"标签名称"		minlength(3) maxlength(100)
//	@Param		state			body		int			false		"状态"			Enums(0, 1)	default(1)
//	@Param		modified_by		body		string		false		"修改者"		minlength(3) maxlength(100)
//	@Success	200		{object}		model.TagSwagger	"成功"
//	@Failure	400		{object}		errcode.Error		"请求错误"
//	@Failure	500		{object}		errcode.Error		"内部错误"
//	@Router		/api/v1/tags/{id}	[put]
func (t Tag) Update (c *gin.Context){}

//	@Summary 	删除标签
//	@Produce 	json
//	@Param		id				body		int			false		"标签ID"
//	@Success	200		{object}		model.TagSwagger	"成功"
//	@Failure	400		{object}		errcode.Error		"请求错误"
//	@Failure	500		{object}		errcode.Error		"内部错误"
//	@Router		/api/v1/tags/{id}	[delete]
func (t Tag) Delete (c *gin.Context){}